+++ include/mpi/mpi_dummy.hpp
@@ -35,6 +35,7 @@ OMPI_DECLSPEC extern struct ompi_predefined_op_t ompi_mpi_op_replace;
 OMPI_DECLSPEC extern struct ompi_predefined_op_t ompi_mpi_op_no_op;
 
 using MPI_Status = int;
+using MPI_Request = int;
 // typedef struct ompi_status_public_t MPI_Status;
 typedef struct ompi_communicator_t *MPI_Comm;
 typedef struct ompi_datatype_t *MPI_Datatype;
@@ -43,7 +44,6 @@ typedef struct ompi_file_t *MPI_File;
 typedef struct ompi_group_t *MPI_Group;
 typedef struct ompi_info_t *MPI_Info;
 typedef struct ompi_op_t *MPI_Op;
-typedef struct ompi_request_t *MPI_Request;
 typedef struct ompi_message_t *MPI_Message;
 typedef struct ompi_win_t *MPI_Win;
 typedef struct mca_base_var_enum_t *MPI_T_enum;